backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
celery>=5.4.0
redis>=5.0.0
deepgram-sdk>=3.7.0
anthropic>=0.40.0
python-docx>=1.1.0
weasyprint>=62.0
python-Levenshtein>=0.25.0
jellyfish>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
jiwer>=3.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
pydub>=0.25.0
ffmpeg-python>=0.2.0
noisereduce>=3.0.0
